Correct field type from double to decimal in "Databases Advanced - Entity Framework/Entity Framework Code First/7. Gringotts Database/Models/WizardDeposit.cs"

+++ Databases Advanced - Entity Framework/Entity Framework Code First/7. Gringotts Database/Models/WizardDeposit.cs
@@ -75,7 +75,7 @@
 
         public double? DepositInterest { get; set; }
 
-        public double? DepositCharge { get; set; }
+        public decimal? DepositCharge { get; set; }
 
         public DateTime? DepositExpirationDate { get; set; }
